Refactor Program.cs to separate classes
