+++ src/wig/Extensions/ProgressExtensions.cs
@@ -1,9 +1,6 @@
 ﻿namespace wig
 {
     using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
     using System.Threading.Tasks;
     using Spectre.Console;
